Python 3.10 do not like module/function name crash, renaming apidata.py
